examples/example_plugin/requirements.txt
# Example dependencies that will be automatically isolated
numpy==1.24.0
pillow==10.0.0
requests==2.31.0